Changing how fractions are written
I realized that we would run into an issue with how to fraction displaying worked before. if you had 1/12 for example it would treat it like a 2 diit fraction but the numerator would only be 1 digit and therefore your numerator wouldnt be centered. now it works by having 3 functions that display a diffrent number of digits and based upon the digits of numerator and denominator putting them in the function that displays that many digits