Multi-Digit Numerator bug fix and Frac line changes
There was a bug in displaying 2 and 3 digit numerators that was caused by having it also try to display it with one digit centering. I also changed FracLine display so that way it displays one contionous line instead of broken up dashes. this was done by overlaying frac line 1 and 2 for a continous frac line 2 and for frac line 3 by overlaying a frac line 2 and 3

Lastly